Move tests for `setItems` into separate file
