+++ test/operations-unit.js
@@ -21,33 +21,6 @@ describe('operations', () => {
     expectExactlySameMembers(getSelection(state), [])
   })
 
-  describe('updating the list of selectable items', () => {
-    it('should remove items from selection that don\'t exist anymore', () => {
-      const state = init()
-
-      const newState = flow(
-        setItems(iterable(['A', 'B', 'C', 'D'])),
-        setSelection(['B', 'C']),
-        setItems(iterable(['A', 'C', 'D']))
-      )(state)
-
-      expectExactlySameMembers(getSelection(newState), ['C'])
-    })
-
-    it('should remove item from anchor that doesn\'t exist anymore', () => {
-      const state = init()
-
-      const newState = flow(
-        setItems(iterable(['A', 'B', 'C', 'D'])),
-        replace('C'),
-        setItems(iterable(['A', 'B', 'D'])),
-        rangeTo('D')
-      )(state)
-
-      expectExactlySameMembers(getSelection(newState), ['A', 'B', 'D'])
-    })
-  })
-
   describe('replace selection with single item', () => {
     it('should replace empty selection with single item', () => {
       const state = init()
